+++ packages/core/src/verification/hash-verifier.ts
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 import { pLimit } from 'plimit-lit';
-import { DataStream, VerificationStrategy } from '../../types/wayfinder.js';
+import { defaultLogger } from '../logger.js';
+import type { DataStream, Logger, VerificationStrategy } from '../types.js';
 import { arioGatewayHeaders } from '../utils/ario.js';
+import { sandboxFromId } from '../utils/base64.js';
 import { hashDataStreamToB64Url } from '../utils/hash.js';
-import { Logger, defaultLogger, sandboxFromId } from '../wayfinder.js';
 
 export class HashVerificationStrategy implements VerificationStrategy {
   private readonly trustedGateways: URL[];