+++ packages/core/src/verification/signature-verifier.ts
@@ -23,19 +23,20 @@ import {
 } from '@dha-team/arbundles';
 
 import { pLimit } from 'plimit-lit';
-import {
+import { GqlClassifier } from '../classifiers/gql-classifier.js';
+import { defaultLogger } from '../logger.js';
+import type {
   DataClassifier,
   DataStream,
+  Logger,
   VerificationStrategy,
-} from '../../types/wayfinder.js';
-import { GqlClassifier } from '../classifiers/gql-classifier.js';
+} from '../types.js';
 import { arioGatewayHeaders } from '../utils/ario.js';
 import { fromB64Url } from '../utils/base64.js';
 import {
   isAsyncIterable,
   readableStreamToAsyncIterable,
 } from '../utils/hash.js';
-import { Logger, defaultLogger } from '../wayfinder.js';
 import { convertDataStreamToDataRoot } from './data-root-verifier.js';
 
 /**